fixing errors introduced by refactoring
