+++ quickfixj-from-fix-orchestra-parent/quickfixj-from-fix-orchestra-codegen-maven-plugin/src/it/settings.xml
