+++ quickfixj-from-fix-orchestra-parent/quickfixj-from-fix-orchestra-codegen-maven-plugin/src/it/simple-it/pom.xml
