+++ quickfixj-from-fix-orchestra-parent/quickfixj-from-fix-orchestra-dictionary-maven-plugin/pom.xml
@@ -25,7 +25,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.quickfixj</groupId>
-			<artifactId>repository-to-quickfixj</artifactId>
+			<artifactId>quickfixj-from-fix-orchestra</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
