+++ quickfixj-from-fix-orchestra-parent/quickfixj-from-fix-orchestra-dictionary-maven-plugin/src/it/simple-it/pom.xml
