+++ quickfixj-from-fix-orchestra-parent/quickfixj-from-fix-orchestra-dictionary-maven-plugin/src/main/java/org/quickfixj/DictionaryGeneratorMojo.java
