+++ quickfixj-from-fix-orchestra-parent/quickfixj-from-fix-orchestra-dictionary-maven-plugin/src/test/java/org/quickfixj/DictionaryGeneratorMojoTest.java
